Rename controller package to manager
The name of the package makes more sense to be manager and not
controller.
