+++ cmd/internal/manager/main.go
@@ -1,4 +1,4 @@
-package controller
+package manager
 
 import (
 	"context"
